requirements.txt
streamlit>=1.35.0
groq>=0.9.0
psycopg2-binary>=2.9.9
pandas>=2.0.0
tabulate>=0.9.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
openpyxl>=3.1.0